@@ -12,15 +12,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.6",
-    "@angular/common": "^4.4.6",
-    "@angular/compiler": "^4.4.6",
-    "@angular/core": "^4.4.6",
-    "@angular/forms": "^4.4.6",
-    "@angular/http": "^4.4.6",
-    "@angular/platform-browser": "^4.4.6",
-    "@angular/platform-browser-dynamic": "^4.4.6",
-    "@angular/router": "^4.4.6",
+    "@angular/animations": "^5.1.1",
+    "@angular/common": "^5.1.1",
+    "@angular/compiler": "^5.1.1",
+    "@angular/core": "^5.1.1",
+    "@angular/forms": "^5.1.1",
+    "@angular/http": "^5.1.1",
+    "@angular/platform-browser": "^5.1.1",
+    "@angular/platform-browser-dynamic": "^5.1.1",
+    "@angular/platform-server": "^5.1.1",
+    "@angular/router": "^5.1.1",
     "bootstrap": "^4.0.0-beta.2",
     "font-awesome": "^4.7.0",
     "helmet": "^3.9.0",
@@ -29,19 +30,19 @@
     "moment": "^2.19.2",
     "ngx-toastr": "^6.5.0",
     "popper.js": "^1.13.0",
-    "rxjs": "^5.5.2",
+    "rxjs": "^5.5.5",
     "sjcl": "^1.0.7",
     "tether": "^1.4.2",
+    "typescript": "2.4.2",
     "zone.js": "^0.8.14"
   },
   "devDependencies": {
     "@angular/cli": "^1.5.3",
-    "@angular/compiler-cli": "^4.4.6",
-    "@angular/language-service": "^4.4.6",
+    "@angular/compiler-cli": "^5.1.1",
+    "@angular/language-service": "^5.1.1",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "^6.0.92",
-    "codelyzer": "~3.0.1",
     "jasmine-core": "~2.6.2",
     "jasmine-spec-reporter": "~4.1.0",
     "karma": "~1.7.0",
@@ -53,7 +54,6 @@
     "protractor": "~5.1.2",
     "ts-node": "~3.0.4",
     "tslint": "~5.3.2",
-    "typescript": "~2.3.3",
     "webpack": "^3.8.1"
   }
 }
