@@ -1,16 +1,14 @@
 import { Injectable } from '@angular/core';
-import {Observable} from 'rxjs/Observable';
-import {Timeline} from '../models/timeline';
-import {WordsHttpService} from './words-http.service';
+import {HttpClient} from "@angular/common/http";
 
 @Injectable()
 export class TimelineService {
 
   constructor(
-    private http: WordsHttpService
+    private http: HttpClient
   ) { }
 
   getTimelineData(month) {
-    return this.http.get('/api/timeline/' + month).map(res => res.json());
+    return this.http.get('/api/timeline/' + month);
   }
 }
