@@ -1,6 +1,5 @@
 import { Injectable } from '@angular/core';
-import {WordsHttpService} from "./words-http.service";
-import {RequestOptions, Headers} from "@angular/http";
+import {HttpClient} from "@angular/common/http";
 
 
 @Injectable()
@@ -12,27 +11,19 @@ export class UserService {
   currentUrl = '/api/me';
 
   constructor(
-    private http: WordsHttpService
+    private http: HttpClient
   ) { }
 
   getCurrentUser() {
-    return this.http.get(this.currentUrl).map(res => res.json());
+    return this.http.get(this.currentUrl);
   }
 
   signup(credentials) {
-    const headers = new Headers({'Content-Type': 'application/json'});
-    const options = new RequestOptions({headers: headers});
-
-    return this.http.post(this.signupUrl, JSON.stringify({username: credentials.username, password: credentials.password}), options)
-      .map(res => res.json());
+    return this.http.post(this.signupUrl, {username: credentials.username, password: credentials.password});
   }
 
   signin(credentials) {
-    const headers = new Headers({'Content-Type': 'application/json'});
-    const options = new RequestOptions({headers: headers});
-
-    return this.http.post(this.signinUrl, JSON.stringify({username: credentials.username, password: credentials.password}), options)
-      .map(res => res.json());
+    return this.http.post(this.signinUrl, {username: credentials.username, password: credentials.password});
   }
 
   logout() {
