@@ -1,31 +1,23 @@
 import {Injectable} from '@angular/core';
-import {Http, Request, Response, RequestOptions, XHRBackend, RequestOptionsArgs} from '@angular/http';
 import {Router} from '@angular/router';
 import {Observable} from 'rxjs/Observable';
+import {HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest} from "@angular/common/http";
 
 @Injectable()
-export class WordsHttpService extends Http {
+export class AuthInterceptor implements HttpInterceptor {
+  constructor(private router: Router) {}
 
-  constructor(
-    backend: XHRBackend,
-    defaultOptions: RequestOptions,
-    private router: Router
-  ) {
-    super(backend, defaultOptions);
-  }
+  intercept(req: HttpRequest<any>, next: HttpHandler): Observable<HttpEvent<any>> {
+    return next.handle(req).catch(err => {
+      if (err instanceof HttpErrorResponse) {
+        if (err.status === 401) {
+          this.router.navigate(['/signin']);
+        }
+        if (err.status === 500) {
 
-  request(url: string | Request, options?: RequestOptionsArgs): Observable<Response> {
-    return super.request(url, options).catch((e) => {
-      if (e.status === 401) {
-        this.router.navigate(['/signin']);
+        }
+        return Observable.throw(err);
       }
-
-      if (e.status === 500) {
-
-      }
-
-      return Observable.throw(e);
-    });
+    })
   }
-
 }
